requirements.txt
anyio==3.6.2
certifi==2022.12.7
click==8.1.3
Flask==2.2.2
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
python-telegram-bot==20.0
rfc3986==1.5.0
sniffio==1.3.0
Werkzeug==2.2.2
zipp==3.11.0
